tensorpool update
