+++ framework/include/nndeploy/inference/default/default_inference_param.h
@@ -26,7 +26,7 @@ class DefaultInferenceParam : public InferenceParam {
 
   ir::ModelDesc *model_desc_ = nullptr;
   net::TensorPoolType tensor_pool_type_ =
-      net::kTensorPool1DSharedObjectTypeGreedyBySizeImprove;
+      net::kTensorPool1DOffsetCalculateTypeGreedyByBreadth;
 };
 
 }  // namespace inference
